Server: update java-telegram-bot-api to version 1.2.2

+++ modules/example-module/src/main/java/com/github/pixelase/bot/modules/example/ExampleUserTask.java
@@ -21,7 +21,7 @@ public class ExampleUserTask extends UserTask {
 			sleep(userTaskDelay);
 			Random random = new Random();
 			if (isMessageUpdated()) {
-				bot.sendMessage(currentMessage.chat().id().intValue(),
+				bot.sendMessage(currentMessage.chat().id(),
 						Emoji.values()[random.nextInt(Emoji.values().length - 1)].toString(), null, null, null, null);
 				System.out.printf("From %s task(%s): %s - %s\n", user.username(), this.hashCode(), currentMessage.text(), currentMessage.messageId());
 			}
